Rename Tests.XUnit -> XUnit.Tests
To fit the existing pattern
