+++ src/EventStore.Core.XUnit.Tests/EventStore.Core.XUnit.Tests.csproj
