+++ src/EventStore.Core.XUnit.Tests/LogAbstraction/IdentityLowHasherTests.cs
@@ -2,15 +2,15 @@
 using Xunit;
 using LogV3StreamId = System.UInt32;
 
-namespace EventStore.Core.Tests.XUnit.LogAbstraction {
-	public class IdentityHighHasherTests {
+namespace EventStore.Core.XUnit.Tests.LogAbstraction {
+	public class IdentityLowHasherTests {
 		[Theory]
 		[InlineData(0, 0)]
-		[InlineData(5, 0)]
-		[InlineData(0xAAAA_BBBB, 0)]
-		[InlineData(0xFFFF_FFFF, 0)]
+		[InlineData(5, 5)]
+		[InlineData(0xAAAA_BBBB, 0xAAAA_BBBB)]
+		[InlineData(0xFFFF_FFFF, 0xFFFF_FFFF)]
 		public void hashes_correctly(LogV3StreamId x, uint expected) {
-			var sut = new IdentityHighHasher();
+			var sut = new IdentityLowHasher();
 			Assert.Equal(expected, sut.Hash(x));
 		}
 	}
