+++ src/EventStore.Core.XUnit.Tests/LogV3/LogV3SizerTests.cs
@@ -1,7 +1,7 @@
 ﻿using EventStore.Core.LogV3;
 using Xunit;
 
-namespace EventStore.Core.Tests.XUnit.LogV3 {
+namespace EventStore.Core.XUnit.Tests.LogV3 {
 	public class LogV3SizerTests {
 		[Fact]
 		public void can_get_size() {
