+++ src/EventStore.Core.XUnit.Tests/LogV3/LogV3StreamIdValidatorTests.cs
@@ -2,7 +2,7 @@
 using EventStore.Core.LogV3;
 using Xunit;
 
-namespace EventStore.Core.Tests.XUnit.LogV3 {
+namespace EventStore.Core.XUnit.Tests.LogV3 {
 	public class LogV3StreamIdValidatorTests {
 		readonly LogV3StreamIdValidator _sut = new();
